feat(oasis): UE 5.7 migration — build, UHT, includes, Blueprints + Levels
UE5 project upgraded from 5.4 to 5.7 and made fully compilable.

## Project & Build System
- `.uproject`, `*.Target.cs`: updated EngineAssociation + paths to UE_5.7
- `ZionOasis.Build.cs`: removed missing `CommonGame`, added `CommonUI` + `SignificanceManager`, renamed `ChaosPhysics` → `Chaos`
- `ZionOasisEditor.Target.cs`: added `bOverrideBuildEnvironment = true`
- `DefaultEngine.ini`: updated renderer settings for UE 5.7 (Lumen, Nanite, D3D12)
- `RunEditor.ps1`: fixed reference from 5.4 → 5.7

## UHT / Codegen Fixes
- `GoldenEggManager.h`: replaced C++14 digit separators (`1'000...`) with plain integers
- `ConsciousnessTypes.h`: added `None = 0` to `EConsciousnessLevel` (UHT requires zero entry)
- Deleted `UE5_AvatarTypes.h` — removed duplicate `EAvatarID` enum

## C++ Compilation Fixes
- All `.h` / `.cpp`: fixed local cross-directory includes to use `ZionOasis/<Dir>/<File>.h` format; same-dir includes use bare filename; engine headers unchanged
- `ZionGameInstance.h/.cpp`: removed `UPROPERTY(Config)` initialization in constructor (invalid for non-static Config properties)
- `ZionOasisGameMode.cpp`: fixed `BroadcastBlockMined` signature (`int64` → `int32` to match header)
- `ZionPlayerController.h/.cpp`: added declared constructor, implemented `BeginPlay` / `OnPossess`, fixed delegate `Broadcast()` with no params, fixed `FRotator&` reference for `GetPlayerViewPoint`
- `ZionCharacter.cpp`: fixed `GetLifetimeReplicatedProps` parameter name (`Out` → `OutLifetimeProps`)
- `TerritoryManager.cpp`: fixed `FText = FString` assignment → `FText::FromString(Name)`

## Blueprint / Asset Creation (Python)
- `CreateBlueprints.py`: fully rewritten for UE 5.7 Python API:
  - `unreal.load_class(None, ...)` instead of single-arg call
  - `AssetTools.create_asset(...)` instead of non-existent `EditorAssetLibrary.create_asset`
  - `set_editor_property("ParentClass", ...)` fallback for `BlueprintFactory`
  - graceful degradation for `InputActionFactory` / `InputMappingContextFactory` (not exposed in UE 5.7 Python)
- Created 8 assets via script:
  - Blueprints: `BP_ZionOasisGameMode`, `BP_ZionCharacter`, `BP_ZionPlayerController`, `BP_ZionHUD`, `BP_GoldenEggManager`, `BP_TerritoryManager`
  - Levels: `LV_MainMenu`, `LV_World`
- `DefaultEngine.ini` already configured with `GlobalDefaultGameMode`, `GameDefaultMap`, `EditorStartupMap`

+++ V3/L4/oasis/ue5/Source/ZionOasis/Game/ZionOasisGameMode.cpp
@@ -1,10 +1,10 @@
 // Copyright 2026 ZION TerraNova. All Rights Reserved.
-#include "Game/ZionOasisGameMode.h"
-#include "Game/ZionGameInstance.h"
-#include "Player/ZionPlayerController.h"
-#include "Player/ZionCharacter.h"
-#include "Consciousness/ConsciousnessComponent.h"
-#include "Blockchain/ZionBlockchainBridge.h"
+#include "ZionOasisGameMode.h"
+#include "ZionGameInstance.h"
+#include "ZionOasis/Player/ZionPlayerController.h"
+#include "ZionOasis/Player/ZionCharacter.h"
+#include "ZionOasis/Consciousness/ConsciousnessComponent.h"
+#include "ZionOasis/Blockchain/ZionBlockchainBridge.h"
 #include "GameFramework/PlayerState.h"
 #include "Engine/World.h"
 #include "Kismet/GameplayStatics.h"
@@ -45,7 +45,7 @@ APlayerController* AZionOasisGameMode::Login(UPlayer* NewPlayer, ENetRole InRemo
 	return Super::Login(NewPlayer, InRemoteRole, Portal, Options, UniqueId, ErrorMessage);
 }
 
-void AZionOasisGameMode::BroadcastBlockMined(const FString& MinerWallet, int64 BlockHeight)
+void AZionOasisGameMode::BroadcastBlockMined(const FString& MinerWallet, int32 BlockHeight)
 {
 	UE_LOG(LogTemp, Log, TEXT("[ZionOasisGameMode] Block #%lld mined by %s — broadcasting XP"),
 		BlockHeight, *MinerWallet);
